requirements.txt
Django==1.5.1
South==0.8.1
argparse==1.2.1
beautifulsoup4==4.2.0
django-taggit==0.10a1
djangorestframework==2.3.5
requests==1.2.3
wsgiref==0.1.2
